feat: wann3d compiling with lib
