+++ src/ProblemData.h
@@ -2,6 +2,7 @@
 #define PROBLEMDATA_H
 
 #include <iostream>
+#include <unordered_map>
 #include "json.hpp"
 #include <pzfmatrix.h>
 #include <pzvec.h>
@@ -15,7 +16,6 @@ class ProblemData
 {
   struct BoundaryData
   {
-    std::string name; // name of the bc
     int matid;             // bc material ID
     int type;              // bc type 0: direct, 1: neumann
     REAL value;            // bc value
@@ -27,13 +27,14 @@ class ProblemData
     int matid;             // domain material ID
     REAL perm;            // domain permeability
     REAL pOrder;           // polynomial approximation order for flux
-    TPZManVector<BoundaryData, 5> BCs; // vector containing all the bcs info
+    std::unordered_map<std::string, BoundaryData> BCs; // map containing all the bcs info
   };
 
   struct WellboreData : public DomainData
   {
     REAL radius; // domain radius
     REAL length; // domain length
+    REAL excentricity; // domain excentricity
   };
 
   struct ReservoirData : public DomainData
@@ -41,10 +42,12 @@ class ProblemData
     REAL height;
     REAL width;
     REAL length;
+    REAL porosity;
   };
 
   struct FluidData
   {
+    std::string name;
     REAL viscosity;
     REAL density;
   };
@@ -81,15 +84,15 @@ public:
 
   using json = nlohmann::json; // declaration of json class
 
-  int m_Verbose; // verbosity level
-  
   WellboreData m_Wellbore; // wellbore data
   
   ReservoirData m_Reservoir; // reservoir data
-
+  
   FluidData m_Fluid; // fluid data
-
+  
   MeshData m_Mesh; // mesh data
+  
+  int m_VerbosityLevel; // verbosity level
 
 public:
   ProblemData();
