+++ targets/wann3d.cpp
@@ -3,7 +3,8 @@
 #endif
 
 #include <iostream>
-#include "TPZWannTools.h"
+#include "TPZWannGeometryTools.h"
+#include "TPZWannApproxTools.h"
 #include <TPZLinearAnalysis.h>
 #include <TPZSSpStructMatrix.h>  //symmetric sparse matrix storage
 #include <pzstepsolver.h>
@@ -12,6 +13,8 @@
 using namespace std;
 
 const int global_nthread = 32;
+
+void PostProceDataForANN(TPZGeoMesh* gmesh, ProblemData* SimData, std::string& filename, const int npts);
 
 int main() {
   std::cout << "--------- Starting simulation ---------" << std::endl;
@@ -54,573 +57,11 @@ int main() {
 
   std::string anndatafile = "anndata.txt";
   const int npts = 401;
-  PostProceDataForANN(gmesh, SimData, anndatafile, npts);
+  PostProceDataForANN(gmesh, &SimData, anndatafile, npts);
 
   delete cmesh;
   delete gmesh;
   std::cout << "--------- Simulation finished ---------" << std::endl;
-}
-
-TPZGeoMesh*
-ReadMeshFromGmsh(std::string file_name) {
-  // read mesh from gmsh
-  TPZGeoMesh* gmesh;
-  gmesh = new TPZGeoMesh();
-  {
-    TPZGmshReader reader;
-    TPZManVector<std::map<std::string, int>, 4> stringtoint(4);
-    stringtoint[3]["volume_reservoir"] = EDomain;
-
-    stringtoint[2]["surface_wellbore_cylinder"] = ESurfWellCyl;
-    stringtoint[2]["surface_wellbore_heel"] = ESurfHeel;
-    stringtoint[2]["surface_wellbore_toe"] = ESurfToe;
-    stringtoint[2]["surface_farfield"] = EFarField;
-    stringtoint[2]["surface_cap_rock"] = EFarField;
-    
-    stringtoint[1]["curve_wellbore"] = ECurveWell;
-    stringtoint[1]["curve_heel"] = ECurveHeel;
-    stringtoint[1]["curve_toe"] = ECurveToe;
-
-    stringtoint[0]["point_heel"] = EPointHeel;
-    stringtoint[0]["point_toe"] = EPointToe;
-    
-    reader.SetDimNamePhysical(stringtoint);
-    reader.GeometricGmshMesh(file_name, gmesh);
-  }
-
-  return gmesh;
-}
-
-void ModifyGeometricMeshToCylWell(TPZGeoMesh* gmesh) {
-  const REAL cylradius = 0.1;
-  const TPZManVector<REAL,3> cylcenter = {0.,0.,0.}, cylaxis = {1.,0.,0.};
-  int64_t nel = gmesh->NElements();
-  for(int64_t iel = 0; iel < nel ; iel++) {
-    TPZGeoEl* gel = gmesh->Element(iel);
-    if(!gel) continue;
-    if(gel->HasSubElement()) DebugStop();
-    if(gel->MaterialId() != ESurfWellCyl) continue;
-    
-    TPZManVector<int64_t, 4> nodeindices;
-    gel->GetNodeIndices(nodeindices);
-    const int nnodes = gel->NCornerNodes();
-    //Moving the nodes to the cylinder surface
-    TPZManVector<REAL,3> xnode(3,0);
-    for (int in = 0; in < nnodes; in++)
-    {
-      gel->NodePtr(in)->GetCoordinates(xnode);
-      // component of xnode that is orthogonal to cyl axis
-      TPZManVector<REAL, 3> x_orth = xnode - cylcenter;
-      const REAL dax = Dot(x_orth, cylaxis);
-      for (int ix = 0; ix < 3; ix++)
-      {
-        x_orth[ix] -= dax * cylaxis[ix];
-      }
-      const auto normdiff = fabs(Norm(x_orth) - cylradius);
-      if (normdiff > 1e-10)
-      {
-        // Moving the node to the cylinder shell
-        REAL computed_radius = Norm(x_orth);
-        for (int ix = 0; ix < 3; ix++)
-        {
-          xnode[ix] = cylcenter[ix] + (x_orth[ix] / computed_radius) * cylradius + dax * cylaxis[ix];
-        }
-        x_orth = xnode - cylcenter;
-        for (int ix = 0; ix < 3; ix++)
-        {
-          x_orth[ix] -= dax * cylaxis[ix];
-        }
-        REAL new_radius = Norm(x_orth);
-        gel->NodePtr(in)->SetCoord(xnode);
-
-        PZError << __PRETTY_FUNCTION__
-                << "\nNode not on cylinder shell: " << gel->NodePtr(in)->Id()
-                << "\nComputed radius: " << computed_radius
-                << "\nGiven radius: " << cylradius
-                << "\nElement index: " << iel
-                << "\nNew coordinates: " << xnode
-                << "\nNew radius: " << new_radius
-                << std::endl;
-      }
-    }
-
-    TPZChangeEl::ChangeToCylinder(gmesh, iel, cylcenter, cylaxis, cylradius);
-  }
-
-  gmesh->BuildConnectivity();
-
-  nel = gmesh->NElements();
-  for(int64_t iel = 0; iel < nel ; iel++) {
-    TPZGeoEl* gel = gmesh->Element(iel);
-    if(!gel) continue;
-    if(gel->HasSubElement()) DebugStop();
-    // if(gel->MaterialId() != ESurfWellCyl && gel->MaterialId() != ESurfHeel && gel->MaterialId() != ESurfToe) continue;
-    if(gel->MaterialId() != ESurfWellCyl) continue;
-    int nsides = gel->NSides();
-    // for (int iside = gel->NCornerNodes(); iside < nsides; iside++) {
-    for (int iside = gel->FirstSide(1); iside < nsides; iside++) {
-      TPZGeoElSide gelSide(gel,iside);
-      TPZStack<TPZGeoElSide> allneigh;
-      for(auto neigh = gelSide.Neighbour(); neigh != gelSide ; neigh++) {      
-        if(neigh.Element()->IsGeoBlendEl()) {
-          continue;
-        }
-        if (neigh.Element()->MaterialId() == ESurfWellCyl) {
-          continue;
-        }
-        // if(neigh.Element()->Dimension() < 2) continue;
-        // if(neigh.Element()->MaterialId() != EDomain) DebugStop();
-        allneigh.Push(neigh);
-      }
-    //   std::cout << "Element " << iel << " side " << iside << " has " << allneigh.size() << " neighbours" << std::endl;
-      for(auto it : allneigh){
-        TPZChangeEl::ChangeToGeoBlend(gmesh, it.Element()->Index());
-      }
-    }
-  }  
-}
-
-void InsertXCoorInSet(const REAL x, std::set<REAL>& nodeCoordsX, const REAL tol) {
-  if (nodeCoordsX.size() == 0) {
-    nodeCoordsX.insert(x);
-    return;
-  }
-  auto it = std::lower_bound(nodeCoordsX.begin(), nodeCoordsX.end(), x);
-  if (it == nodeCoordsX.end()) {
-    REAL ref = *(--it);
-    if (fabs(x - ref) > tol) {
-      nodeCoordsX.insert(x);
-    }
-  } else {
-    REAL ref1 = *it;
-    if (fabs(x - ref1) <= tol) {
-      return;
-    }
-    REAL ref2 = *(--it);
-    if (fabs(x - ref2) <= tol) {
-      return;
-    }
-    nodeCoordsX.insert(x);
-  }
-}
-
-REAL FindClosestX(const REAL x, const std::set<REAL>& nodeCoordsX, const REAL tol) {
-  REAL closestX = -1000;
-  auto it = std::lower_bound(nodeCoordsX.begin(), nodeCoordsX.end(), x);
-  if (it == nodeCoordsX.end()) {
-    closestX = *(--it);
-    if(fabs(x - closestX) >= tol) DebugStop();
-    return closestX;
-  }
-  REAL ref = *it;
-  if (fabs(x - ref) <= tol) {
-    closestX = ref;
-    return closestX;
-  }
-  ref = *(--it);
-  if (fabs(x - ref) <= tol) {
-    closestX = ref;
-    return closestX;
-  }
-  DebugStop();
-  return -1;
-}
-
-void CreatePressure2DElsAndOrderIds(TPZGeoMesh* gmesh) {
-  const int nel = gmesh->NElements();
-  const REAL tol = 1.e-4;
-  std::set<int64_t> pressure2Dels;
-  std::set<REAL> nodeCoordsX;
-  for (int64_t iel = 0; iel < nel; iel++) {
-    TPZGeoEl* gel = gmesh->Element(iel);
-    if (gel->MaterialId() != ESurfWellCyl) continue;  
-    if (gel->HasSubElement()) continue;
-    pressure2Dels.insert(iel);
-    TPZGeoElBC bc(gel,gel->NSides()-1,EPressure2DSkin); 
-    TPZGeoElBC bc2(gel,gel->NSides()-1,EPressureInterface);
-    TPZGeoElBC bc3(gel,gel->NSides()-1,EHDivBoundInterface);
-    for (int i = 0; i < gel->NCornerNodes(); i++) {
-      TPZManVector<REAL,3> coor(3);
-      gel->NodePtr(i)->GetCoordinates(coor);
-      InsertXCoorInSet(coor[0], nodeCoordsX, tol);
-    }
-  }
-
-  std::map<REAL,std::set<int64_t>> xToNodes;
-  for (auto iel : pressure2Dels) {
-    TPZGeoEl* gel = gmesh->Element(iel);
-    if (gel->MaterialId() != ESurfWellCyl) DebugStop();
-    for (int i = 0; i < gel->NCornerNodes(); i++) {
-      TPZManVector<REAL,3> coor(3);
-      gel->NodePtr(i)->GetCoordinates(coor);
-      REAL closestX = FindClosestX(coor[0], nodeCoordsX, tol);
-      xToNodes[closestX].insert(gel->NodeIndex(i));
-    }
-  }
-
-  int64_t maxid = gmesh->CreateUniqueNodeId();
-  for (auto& it : xToNodes) {
-    const REAL x = it.first;
-    const auto& nodes = it.second;
-    const int64_t nnodes = nodes.size();
-    if (nnodes < 2) DebugStop();          
-    for (auto& node : nodes) {
-      gmesh->NodeVec()[node].SetNodeId(maxid++);
-    }
-  }
-
-  // Print nodes for debugging
-  // std::multimap<REAL,int64_t> xToNodeIds;
-  // for (auto& it : xToNodes) {    
-  //   for (auto& node : it.second) {
-  //     const int64_t id = gmesh->NodeVec()[node].Id();
-  //     const REAL x = gmesh->NodeVec()[node].Coord(0);
-  //     xToNodeIds.insert({x,id});
-  //   }        
-  // }  
-  // for (const auto& it : xToNodeIds) {
-  //   std::cout << "X: " << it.first << " Node ID: " << it.second << std::endl;
-  // }
-}
-
-void AddPressureSkinElements(TPZCompMesh* cmesh, const int pordWell, const int laglevel) {
-  const int dim = cmesh->Dimension()-1; // 2D for these pressure elements living in the boundary of the 3d well
-  const int matid = EPressure2DSkin;
-  TPZNullMaterial<STATE>* mat = new TPZNullMaterial<>(matid,dim);
-  cmesh->SetAllCreateFunctionsContinuous();
-  cmesh->ApproxSpace().CreateDisconnectedElements(true);
-  cmesh->InsertMaterialObject(mat);
-  cmesh->SetDefaultOrder(pordWell);
-
-  std::set<int> matidset = {matid};
-  cmesh->AutoBuild(matidset);
-
-  const bool isContinuousConnects = false; // the pressure skin elements should not be continuous if HDiv elements in the wellbore.
-  if(isContinuousConnects){ // This leads to continuous elements
-    std::set<REAL> nodeCoordsX;
-    const int64_t nel = cmesh->NElements();
-    std::set<int64_t> pressure2Dels;
-    for (int64_t iel = 0; iel < nel; iel++) {
-      TPZCompEl* cel = cmesh->Element(iel);    
-      if (!cel) continue;
-      if (cel->Material()->Id() != matid) continue;
-      TPZGeoEl* gel = cel->Reference();
-      if(gel->HasSubElement()) continue;
-      if(gel->NNodes() != 4) DebugStop();
-      TPZInterpolatedElement* intEl = dynamic_cast<TPZInterpolatedElement*>(cel);
-      if (!intEl) DebugStop();    
-      // I am at a pressure 2D element at the 3d well boundary
-      pressure2Dels.insert(iel);
-      for(int i = 4 ; i < 8 ; i++){
-        TPZConnect& c = cel->Connect(i);
-        REAL x0 = gel->NodePtr(i%4)->Coord(0), x1 = gel->NodePtr((i+1)%4)->Coord(0);
-        InsertXCoorInSet(x0, nodeCoordsX, 1.e-6);
-        InsertXCoorInSet(x1, nodeCoordsX, 1.e-6);
-        InsertXCoorInSet((x0+x1)/2., nodeCoordsX, 1.e-6);
-        if (fabs(x0 - x1) < 1.e-6) {
-          // Both nodes are on the same x coordinate
-          c.SetOrder(1);
-          c.SetNShape(0);
-        }
-      }
-
-      // Setting the area connect to order 1
-      TPZConnect& c = cel->Connect(8);
-      c.SetOrder(1);
-      c.SetNShape(0);
-
-      for (int i = 0; i < 9; i++) {
-        TPZConnect& c = cel->Connect(i);
-        c.SetLagrangeMultiplier(laglevel);
-      }
-      
-    }
-
-    std::map<REAL,std::set<int64_t>> xToConnects;
-    const REAL tol = 1.e-6;
-    for (auto iel : pressure2Dels) {
-      TPZCompEl* cel = cmesh->Element(iel);
-      TPZGeoEl* gel = cel->Reference();
-      if(gel->HasSubElement()) DebugStop();
-      for (int i = 0; i < gel->NCornerNodes(); i++) {
-        TPZManVector<REAL,3> coor(3);
-        gel->NodePtr(i)->GetCoordinates(coor);
-        REAL closestX = FindClosestX(coor[0], nodeCoordsX, tol);
-        xToConnects[closestX].insert(cel->ConnectIndex(i));
-      }
-      for (int i = 4; i < 8; i++) {
-        REAL x0 = gel->NodePtr(i%4)->Coord(0), x1 = gel->NodePtr((i+1)%4)->Coord(0);
-        if (fabs(x0 - x1) < 1.e-6) {
-          continue;
-        }
-        REAL closestX = FindClosestX((x0+x1)/2., nodeCoordsX, tol);
-        xToConnects[closestX].insert(cel->ConnectIndex(i));
-      }    
-    }
-    
-    for (auto iel : pressure2Dels) {
-      TPZCompEl* cel = cmesh->Element(iel);
-      TPZGeoEl* gel = cel->Reference();
-
-      for (int i = 0; i < gel->NCornerNodes(); i++) {
-        TPZManVector<REAL,3> coor(3);
-        gel->NodePtr(i)->GetCoordinates(coor);
-        REAL closestX = FindClosestX(coor[0], nodeCoordsX, tol);
-        int64_t cindex = *xToConnects[closestX].begin();
-        cel->SetConnectIndex(i,cindex);
-      }
-      for (int i = 4; i < 8; i++) {
-        REAL x0 = gel->NodePtr(i%4)->Coord(0), x1 = gel->NodePtr((i+1)%4)->Coord(0);
-        if (fabs(x0 - x1) < 1.e-6) {
-          continue;
-        }
-        REAL closestX = FindClosestX((x0+x1)/2., nodeCoordsX, tol);
-        int64_t cindex = *xToConnects[closestX].begin();
-        cel->SetConnectIndex(i,cindex);
-      }    
-    }
-  }
-  else {    
-    const int64_t nel = cmesh->NElements();
-    std::set<int64_t> pressure2Dels;
-    std::map<REAL,TPZManVector<REAL,3>> elCentToConnects;
-    std::set<REAL> elCentX;
-    const REAL tol = 1.e-6;
-    for (int64_t iel = 0; iel < nel; iel++) {
-      TPZCompEl* cel = cmesh->Element(iel);    
-      if (!cel) continue;
-      if (cel->Material()->Id() != matid) continue;
-      TPZGeoEl* gel = cel->Reference();
-      if(gel->HasSubElement()) DebugStop();
-      if(gel->NNodes() != 4) DebugStop();
-
-      TPZGeoElSide gelside(gel);
-      TPZManVector<REAL,3> cent(3);
-      gelside.CenterX(cent);
-      InsertXCoorInSet(cent[0], elCentX, 1.e-6);
-      REAL closestX = FindClosestX(cent[0], elCentX, tol);      
-      if (elCentToConnects.find(closestX) == elCentToConnects.end()) {
-        TPZManVector<REAL> newconnects(3);
-        newconnects[0] = cmesh->AllocateNewConnect(1, 1, pordWell);
-        newconnects[1] = cmesh->AllocateNewConnect(1, 1, pordWell);
-        newconnects[2] = cmesh->AllocateNewConnect(pordWell-1, 1, pordWell);
-        elCentToConnects[closestX] = newconnects;        
-      }
-
-
-      TPZInterpolatedElement* intEl = dynamic_cast<TPZInterpolatedElement*>(cel);
-      if (!intEl) DebugStop();    
-      // I am at a pressure 2D element at the 3d well boundary
-      pressure2Dels.insert(iel);
-      for(int i = 4 ; i < 8 ; i++){
-        TPZConnect& c = cel->Connect(i);
-        REAL x0 = gel->NodePtr(i%4)->Coord(0), x1 = gel->NodePtr((i+1)%4)->Coord(0);
-        if (fabs(x0 - x1) < 1.e-6) {
-          // Both nodes are on the same x coordinate
-          c.SetOrder(1);
-          c.SetNShape(0);
-        }
-      }
-
-      // Setting the area connect to order 1
-      TPZConnect& c = cel->Connect(8);
-      c.SetOrder(1);
-      c.SetNShape(0);
-
-      for (int i = 0; i < 9; i++) {
-        TPZConnect& c = cel->Connect(i);
-        c.SetLagrangeMultiplier(laglevel);
-      }      
-    }
-
-    for (auto iel : pressure2Dels) {
-      TPZCompEl* cel = cmesh->Element(iel);
-      TPZGeoEl* gel = cel->Reference();
-      TPZGeoElSide gelside(gel);
-      TPZManVector<REAL,3> cent(3);
-      gelside.CenterX(cent);
-      REAL closestX = FindClosestX(cent[0], elCentX, tol);
-      auto it = elCentToConnects.find(closestX);
-      TPZManVector<REAL,3> newconnects;
-      if (it != elCentToConnects.end()) {
-          newconnects = it->second;
-      } else {
-          DebugStop(); // closestX not found in elCentToConnects
-      }
-
-      for (int i = 0; i < gel->NCornerNodes(); i++) {
-        TPZManVector<REAL,3> coor(3);
-        gel->NodePtr(i)->GetCoordinates(coor);
-        if (fabs(coor[0] - closestX) < tol) DebugStop();
-        if (coor[0] < closestX) {
-          cel->SetConnectIndex(i,newconnects[0]);
-        } else {
-          cel->SetConnectIndex(i,newconnects[1]);
-        }
-      }
-      for (int i = 4; i < 8; i++) {
-        REAL x0 = gel->NodePtr(i%4)->Coord(0), x1 = gel->NodePtr((i+1)%4)->Coord(0);
-        if (fabs(x0 - x1) < 1.e-6) {
-          continue;
-        }
-        const REAL midpoint = (x0+x1)/2.;
-        if (fabs(midpoint - closestX) > tol) DebugStop(); // Has to be in the middle!
-        cel->SetConnectIndex(i,newconnects[2]);        
-      }    
-    }
-  }
-
-  cmesh->ComputeNodElCon();
-  cmesh->CleanUpUnconnectedNodes();  
-  for (int64_t i = 0; i < cmesh->NConnects(); i++) {
-    TPZConnect &c = cmesh->ConnectVec()[i];
-    if (c.NElConnected() == 0) continue;
-    cmesh->Block().Set(c.SequenceNumber(),c.NShape()*c.NState());
-  }
-  cmesh->InitializeBlock();
-
-  {
-    std::ofstream out("cmesh.txt");
-    cmesh->Print(out);
-  }
-}
-
-void AddWellboreElements(TPZVec<TPZCompMesh*>& meshvec, const int pordWell, const int laglevel) {
-  const int dimwell = 1;
-  const int matid = ECurveWell;
-
-  // First create the pressure elements
-  TPZNullMaterial<STATE>* mat = new TPZNullMaterial<>(matid,dimwell);
-  meshvec[1]->SetAllCreateFunctionsContinuous();
-  meshvec[1]->ApproxSpace().CreateDisconnectedElements(true);
-  meshvec[1]->InsertMaterialObject(mat);
-  meshvec[1]->SetDefaultOrder(pordWell);
-
-  std::set<int> matidset = {matid};
-  meshvec[1]->AutoBuild(matidset);
-
-  const int64_t nel = meshvec[1]->NElements();
-  for (int64_t iel = 0; iel < nel; iel++) {
-    TPZCompEl* cel = meshvec[1]->Element(iel);
-    if (!cel) continue;
-    if (cel->Material()->Id() != matid) continue;
-    if (cel->Reference()->HasSubElement()) continue;
-    TPZGeoEl* gel = cel->Reference();
-    if(gel->NNodes() != 2) DebugStop();
-    TPZInterpolatedElement* intEl = dynamic_cast<TPZInterpolatedElement*>(cel);
-    if (!intEl) DebugStop();    
-    // I am at a wellbore element
-    for (int i = 0; i < intEl->NConnects(); i++) {
-      TPZConnect& c = cel->Connect(i);
-      c.SetLagrangeMultiplier(laglevel);
-    }
-  }
-
-  // Now create the flux elements in meshvec[0]
-  TPZNullMaterial<STATE>* matFlux = new TPZNullMaterial<>(matid,dimwell);
-  TPZNullMaterial<STATE>* matFluxBcHeel = new TPZNullMaterial<>(EPointHeel,dimwell-1);
-  TPZNullMaterial<STATE>* matFluxBcToe = new TPZNullMaterial<>(EPointToe,dimwell-1);
-  meshvec[0]->InsertMaterialObject(matFlux);
-  meshvec[0]->InsertMaterialObject(matFluxBcHeel);
-  meshvec[0]->InsertMaterialObject(matFluxBcToe);
-
-  meshvec[0]->SetDimModel(dimwell);
-  meshvec[0]->ApproxSpace().SetAllCreateFunctionsHDiv(dimwell);
-  meshvec[0]->SetDefaultOrder(pordWell);
-  
-  std::set<int> matidsetFlux = {matid, EPointHeel, EPointToe};
-  meshvec[0]->AutoBuild(matidsetFlux);
-
-}
-
-void AddHDivBoundInterfaceElements(TPZCompMesh* cmesh, const int porder) {
-  cmesh->Reference()->ResetReference();
-  cmesh->LoadReferences();
-  const int dim = cmesh->Reference()->Dimension();
-  const int matid = EHDivBoundInterface;
-  TPZNullMaterial<STATE>* mat = new TPZNullMaterial<>(matid, dim - 1);
-  cmesh->SetDimModel(dim);
-  cmesh->SetAllCreateFunctionsHDiv();
-  cmesh->InsertMaterialObject(mat);
-  cmesh->SetDefaultOrder(porder);
-
-  std::set<int> matidset = {matid};
-  cmesh->AutoBuild(matidset);
-}
-
-void AddInterfaceElements(TPZMultiphysicsCompMesh* cmesh, const int matidpressure, const int matidinterface, const int laglevel) {
-  cmesh->Reference()->ResetReference();
-  cmesh->LoadReferences();
-  TPZGeoMesh* gmesh = cmesh->Reference();  
-  const int dim = gmesh->Dimension();
-  const int64_t nel = gmesh->NElements();
-  for (int64_t iel = 0; iel < nel; iel++) {
-    TPZGeoEl* gel = gmesh->Element(iel);
-    if (!gel) continue;
-    if (gel->MaterialId() != EPressureInterface) continue;
-    if (gel->HasSubElement()) DebugStop();
-    if (gel->Dimension() != dim-1) DebugStop();
-    TPZCompElSide comp_pressureSide, comp_hdivSide;
-    TPZGeoElSide gelSide(gel);
-    for(auto neigh = gelSide.Neighbour(); neigh != gelSide ; neigh++) {
-      if(neigh.Element()->MaterialId() == EPressure2DSkin) {
-        comp_pressureSide = neigh.Reference();
-      }
-      if(neigh.Element()->MaterialId() == EHDivBoundInterface) {
-        comp_hdivSide = neigh.Reference();
-      }
-    }
-    if (!comp_pressureSide || !comp_hdivSide) DebugStop();
-    TPZMultiphysicsInterfaceElement* interfaceel = new TPZMultiphysicsInterfaceElement(*cmesh,gel,comp_pressureSide,comp_hdivSide);
-  }
-}
-
-void EqualizePressureConnects(TPZCompMesh* cmesh) {
-  cmesh->Reference()->ResetReference();
-  cmesh->LoadReferences();
-  TPZGeoMesh* gmesh = cmesh->Reference();
-  const int dim = gmesh->Dimension();
-  const int64_t nel = gmesh->NElements();
-  for(auto& gel: gmesh->ElementVec()) {
-    if (!gel) continue;
-    if (gel->MaterialId() != ECurveWell) continue;
-    if(gel->HasSubElement()) continue;
-    TPZGeoElSide gelside(gel);
-    TPZGeoElSide surfwellside = gelside.Neighbour();
-    while (surfwellside.Element()->MaterialId() != EPressure2DSkin) {
-      if (surfwellside.Element()->HasSubElement()) DebugStop();
-      if(surfwellside == gelside) DebugStop();      
-      surfwellside = surfwellside.Neighbour();
-    }
-    TPZCompEl* cel = gel->Reference();
-    TPZCompEl* celneigh = surfwellside.Element()->Reference();
-
-    // Set the internal connect of the wellbore as the edge of a 2d pressure skin element
-    const int64_t cindex = celneigh->ConnectIndex(surfwellside.Side());
-    cel->SetConnectIndex(2,cindex);
-
-    // Set the nodal connect of the 1d wellbore elements equal to the 2d pressure skin element nodal connects
-    const int64_t gindex0 = gel->NodeIndex(0), gindex1 = gel->NodeIndex(1);
-    TPZGeoEl* gelsurf = surfwellside.Element();
-    const int64_t nindex0 = gelsurf->SideNodeIndex(surfwellside.Side(),0), nindex1 = gelsurf->SideNodeIndex(surfwellside.Side(),1);
-    const int sidenodelocindex0 = gelsurf->SideNodeLocIndex(surfwellside.Side(),0), sidenodelocindex1 = gelsurf->SideNodeLocIndex(surfwellside.Side(),1);
-    if(nindex0 == gindex0 && nindex1 == gindex1){
-      cel->SetConnectIndex(0,celneigh->ConnectIndex(sidenodelocindex0));
-      cel->SetConnectIndex(1,celneigh->ConnectIndex(sidenodelocindex1));
-    } else if(nindex0 == gindex1 && nindex1 == gindex0){
-      cel->SetConnectIndex(0,celneigh->ConnectIndex(sidenodelocindex1));
-      cel->SetConnectIndex(1,celneigh->ConnectIndex(sidenodelocindex0));
-    } else {
-      DebugStop();
-    }
-
-    std::cout << "Connect indexes of cel: ";
-    for (int i = 0; i < cel->NConnects(); i++) {
-      std::cout << cel->ConnectIndex(i) << " ";
-    }
-    std::cout << std::endl;
-  }
-  cmesh->CleanUpUnconnectedNodes();
 }
 
 void PostProceDataForANN(TPZGeoMesh* gmesh, ProblemData* SimData, std::string& filename, const int npts) {
@@ -647,7 +88,7 @@ void PostProceDataForANN(TPZGeoMesh* gmesh, ProblemData* SimData, std::string& f
     const REAL x = x0 + i * dx;
     TPZManVector<REAL,3> qsi(1, 0.0);    
     xvec[0] = x;
-    TPZGeoEl* gel = TPZGeoMeshTools::FindElementByMatId(gmesh, xvec, qsi, InitialElIndex, {SimData->ECurveWell}, 1.e-6);
+    TPZGeoEl* gel = TPZGeoMeshTools::FindElementByMatId(gmesh, xvec, qsi, InitialElIndex, {SimData->ECurveWell});
     if (!gel) {
       std::cout << "Element not found for x = " << x << std::endl;
       DebugStop();
@@ -674,6 +115,9 @@ void PostProceDataForANN(TPZGeoMesh* gmesh, ProblemData* SimData, std::string& f
     cel->Solution(qsi, qind, output);
     const REAL divq = output[0];
 
+    auto& WellboreData = SimData->m_Wellbore;
+    const REAL wellRad = WellboreData.radius;
+    const REAL pff = WellboreData.BCs["surface_farfield"].value;
     const REAL K = divq / (pff - pressure);
 
     std::cout << "x = " << x << " p = " << pressure << " divq = " << divq << " K = " << K << std::endl;
